backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Werkzeug
itsdangerous
